requirements.txt
anthropic>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
yfinance>=0.2.28
streamlit>=1.28.0
